Fix and expand dead basic block removal.
The specialization graph builder does dead basic block removal before
it computes the dominance tree, to avoid problems in the calculation.
In some cases, however, it was finding and removing basic blocks in
spesh graphs formed from previously optimized code. This was because
the basic block removal was deliberately incomplete, to work around a
lack of moving/removing exception handler regions and being able to
remove inlinees that turned out to be unused.

With these changes, we can now more aggressively remove dead basic
blocks, fixing the issues that cropped up due to work to make it
possible to do more inlining. It's likely these could have been hit
before the current round of spesh work (and were probably part of the
failures with NODELAY going back ages).