More sophisticated named arg handling in spesh.
Previously, we would keep track of them using the bit field used by
unspecialized named argument handling. Since in an inline the use of
that could conflict, the op to do the tracking was `:noinline`. Thus
an attempt was made to not include them if we had no slurpy args and
no deopt points. However, recent spesh changes have introduced more
deopt points, frustrating inlining of things that take named args,
such as `$!decoder.consume-line-chars(:$!chomp)`.

With this change, we no longer need to do any tracking of named args
usage at runtime when in specialized code. Rather, we calculate it all
at spesh time. We keep the calculated bit field around for deopt. A
further notable improvement, required along the way, is that we now
produce code to populate the slurpy hash from the args buffer, which
avoids flag checks to see what should be boxed, and also meaning that
this code can be further lowered by the JIT.