When logging invokes, don't store closures.
This can, in some cases, seriously extend the lifetimes of objects
that were closed over. Instead, just log what we'll really want: the
static frame of the invocation target (which actually simplifies the
handling in the stats code) and whether the caller was its outer (this
will be useful for future optimizations).