ARGLIST on should not be fussy about type
I think a casual user could be forgiven for not caring about the
difference about int and ptr and reg (although if portability is a
concern, it is a real difference). So the code for selecting arg
locations should not be fussy about that difference (the ABI isn't).

This appears to fix the bugs in CORE.setting compilation reported
by nwc10++, but I'm not 100% positive since ASAN on OS X does not
quite meet my expectations.