Eliminate "blocked by alias" propagation checks.
We explicitly guard decont now, and so cannot run into these kinds of
issues any more (and the guards that looked into containers are gone
also). We will do some alias analysis in the future to remove some of
the guards when we can prove they aren't needed. But we are now safe
by default, without needing a check like this.