More precise CFGs for control exception handlers.
Prior to this, they were always linked from the entry block. This
meant that:

1) PHI nodes were always inserted at the start of a handler, and the
   join always included the entry point. Since this has an empty set
   of facts, it conservatively kept us from making assumptions as to
   what types/values would be in place on the exceptional path.
2) The handler could never become unreachable in the CFG, meaning that
   it was never elligible for dead BB elimination.

This change inserts CFG edges to the handler GOTO address from all
basic blocks between the start (inclusive) and end (exclusive) of the
handler. This eliminates the PHIs from the entry node, meaning that
more precise facts may be available for use in optimization. It also
allows for the handler to become unreachable and thus eliminated if
all code that may throw something that would be caught by it would be
eliminated.