Add indexim_s op and improve/fix bugs in eqatim_s
eqatim_s used to only check that the first codepoints of string a and b
were the same using MVM_string_ord_basechar_at. The
MVM_string_equal_at_ignore_mark function was added and we now use this
instead for the eqatim_s op which fixes the bug when it would say the
two strings were eqat when only one codepoint was.

MVM_string_index_ignore_mark was also added, which is accessible as the indexim_s
op.