Implement speculative call optimization.
This is optimization of calls where we can't 100% determine the code
object that will be invoked, but we know from our logging that it is
likely to be consistent. Thus, we guard the static frame that's being
invoked and then can either fast-invoke (pre-selection of the spesh
candidate) or inline.