Correct handler order when inlining.
We duplicate the exception handlers in force when inlining. It's true
that order matters, and innermost first is searched for. Unfortunately
the effort to do the right thing about that was misguided here, since
we're copying from a table that's already sorted, so should just leave
it in the order it was already in.