Implement inlining of basic lexical lookups.
This means that we can now inline closures that look up symbols they
close over. It works by introducing a new pair of lookup ops that find
the code ref in a register and use it to find the outer. After JIT,
this turns out to be as many dereferences as in the non-inlined case.