Make sure the libffi callback can be reused.
Rakudo test t/04-nativecall/21-callback-other-thread.t tests 2..9
fail with a segmentation fault or glibc error (double free or
malloc() problem) when MoarVM is built with --has-libffi.

The problem is due to the callback handler freeing the memory for
its first argument (the libffi call interface, cif).  Doing this
makes the callback not reusable (the callback still exists in
tc->native_callback_cache).