Fix a bug in index/eqat(im) and in ord_getbasechar
Previously it assumed that once we had gotten a synthetic's base character
that we already had the final base character. This wasn’t true for example
with: "\c[LATIN SMALL LETTER J WITH CARON, COMBINING DOT BELOW]"
The "j with caron" would not be decomposed because it was the base character
of a synthetic.

This also fixes a bug in indexim and eqatim ops which caused it to fail when
encountering a synthetic.

We now have ord_getbasechar handle synthetic and non-synthetic's and have
MVM_string_ord_basechar_at just handle the needed string based checks,
then grabbing the grapheme from the string and passing its value on to
ord_getbasechar.