Move thread to starting after running list insert.
There is code to mark the thread contexts of not-yet-started threads
in the MVMThread REPR's gc_mark function. Once it's in starting state
it participates in GC "as normal". Recent refactors changed the way
the insertion into this list is handled, but didn't move the location
of the thread state update. This moves it to happen only after the
thread has been placed into the running list, thus avoid a period when
the thread context would go unmarked.