Put inlined blocks between their caller and its succ
Previously inlined callees were added to the end of the basic block list.
Also, active handlers at the point of the invoke were copied and the
inlinee's code surrounded by annotations pointing at those copies.

Instead we now put the inlined blocks into the list at the position of the
invoke op. We move the inlinee's handlers to the top of the table so they will
be searched first. This means we have to adjust the indexes of the inliner's
annotations but that's actually simpler than the whole copying business.

With this the goto ops for entering and exiting inlined code can be optimized
away.