Don't guardsf results of multi resolutions.
We aren't smart enough to handle this case yet, so it caused a ton
of misses and thus deoptimizations. This fixes a performance loss in
a bunch of programs from invocation speculation (so improves the
Text::CSV benchmark again, probably back to its best), while retaining
the performance gain for things that won from it (such as the million
line reading benchmark).