RT 131375: Fix calculation for maximum array size
We were off by elem_size - log2(elem_size). As we're calculating the
number of bits available for adressing individual array elements, we
have to substract the number of bits for addressing individual bytes in
an array element rather than the number of bytes.
For 64 bit array elements, i.e. 8 bytes, we need 3 bits for addressing
individual bytes. So we lose 3 bits when adressing elements in the
array.