introduce a binary format for heapsnapshot
it starts with a version identifier to differentiate between
the old format and this one. all records except the string heap
are fixed sizes and the file ends in a table outlining the
sizes of the individual segments.

each segment starts in a little identifier to make sure the
snapshot parser doesn't read bogus data, and before a set
of entries there's an entry of the entry size.

the heap analyzer perl6 program can parse this faster and uses
less ram, but the heap dump files themselves are noticably
bigger at the moment.