Make `eqat` 1.7x-2.5x faster
For strings a and b:
1. Both a and b are strands
2. Both a and b are flat
3. One is flat and one is a strand

For these three cases, we have three different codepaths. This allows us
to always use MVM_string_gi_get_grapheme for strands and use
MVM_string_get_grapheme_at_nocheck for flat strings.

The speedup is 1.7x when one is a strand and one is flat, and 2.5x when
both strings are flat.

As before, when both are flat and both are the same number of bits, we use
memcmp.