Rework MVMNFGSynthetic to not store base separately (#657)
This is done to (eventually) get Prepend support working. Unlike
all other marks which come after a base character, Prepend characters
come *before* the base character. All of our current code assumed the
first codepoint of the synthetic is the base character.

For now, we *also* assume the first codepoint of the synthetic is the
base character, but we now have a base_index which will be able to hold
the index of the base character in the codepoint array.

While this doesn't add Prepend support to everything, this is one step
toward getting that working, and decoupling base codepoints from being
the first codepoint of a synthetic.