UCD: For sequences, convert from \x{ } to the Unicode codepoints
This doesn't affect using "\c[ ]" since names with # were not accessible
but it *does* work with .parse-names or with the nqp call directly.

Currently the only thing that uses this is `keycap: #` which uses the \x
syntax so that it parses correctly (though previous versions of the file
had an actual `#` in there, which is why we didn't have this problem before).