make references variable-length, add halfway to index
this reduces the average size of references entries from
24 bytes to something more like 7.6 bytes on average in
one example dataset. The old format took 9.7 chars on
average in the same dataset.

the halfway point in the index serves to allow a second
thread to start parsing half-way into the blob, which
is what allows the new heap snapshot parser to be so fast.