Do dead BB elimination pass after branch deletion
This means that writers in eliminated basic blocks will be marked dead
right away, which in turn means that we'll be able to be more certain
in a bunch of PHI merges. The more precise type information can then
lead to better downstream optimizations.

In a Perl 6 write a million lines benchmark, as well as the immediate
gains this leads to a further inlining, giving 10%-15% time reduction
and bringing the benchmark to within 1.2x of Perl 5.