Don't decont native refs when spesh logging
That will result in the value being boxed. This allocation could make
a problem by causing a GC run to happen, but much more importantly it
leads to rather misleading spesh stats that could in turn end up with
selection of a bogus multi-dispatch candidate during optimization.
This thus fixes the mis-specialization that has caused failures in the
atomics spectests.