Handle native refs better in spesh stats
Previously, we would always consider a type tuple involving one as
being incomplete, since a NativeRef is a container but we don't (any
more) log the type of its content (and when we did, we got it wrong by
logging the boxing of the actually referenced type).

This allows for generating observed type specializations of things
that get passed native references. It's not enough, however, to see
invocations of those be fastinvoke'd or inlined (at least, it not in
the multi-dispatch case).