Set facts on nativeref deconts.
It would, at some point, be nice to have better code-gen for the
decont/box itself. But for now, we can get a notable win from setting
up the facts on the target of the decont. We know the type of the
result and also that it will be concrete, which in turn allows for
full elimination of parameter checks in the case infix:<==>(Int, Int)
is called with an Int and a native reference to an int. That in turn
enables it to be inlined, making this a notable improvement for code
that mixes natives and boxeds (which is decidedly easily done). That
inlining is also a prerequisite for future improvements to get rid of
box/unbox pairs.