Add a reliable locals count for inlines
When we do a multi-level inline, then we won't have a spesh graph for
the nested inline, and so cannot simply ask it for its locals count.
The static frame is also not reliable for this purpose, as the inline
may itself have inlines. Thus just add a field holding the value, so
we reliably have it available.