Move throw -> goto optimization to second pass.
Meaning that we can now apply it to inlined code. Also fix it up to
cope with being used across inlines. This means that Perl 6 next, last,
and redo may potentially be rewritten into gotos (in Perl 6, throws of
control exceptions are always inside of multi candidates, which we've
been able to inline for a while, but could not rewrite to a goto since
the optimization didn't work across inlines until now).