Merge Full Unicode Collation Algorithm Implementation
This is a full implementation of the Unicode Collation Algorithm. We iterate
by codepoint and put this into a ring buffer. The ring buffers hold the exact
number of codepoints which comprise the longest sequence of codepoints which
map to its own collation keys in the Unicode Collation Algorithm. As of Unicode
9.0 this number was 3. When Generate-Collation-Data.p6 is run, this number will
be updated in case future versions contain a higher number. Once the codepoint
iteration into the ringbuffer finds a non-matching codepoint, whether the two
codepoints are less or more than each other is saved in a variable for later in
case we end up needing to break a tie by codepoint.

We only pull as many keys as needed, so usually we only need to use what is in
the ring buffer. Then we compare by primary levels, all the keys pushed so far.
If all the primaries match then we iterate more codepoints and push the
collation values onto a stack. This stack is malloced and can expand as needed,
but this should practically never be the case.

Using the stack how we do is a modified verrsion of the UCA which lets us not
have to push all primaries from start to end of the string onto a one
dimensional array, and then after that push all the secondary, then all the
tertiary.

Codepoints which have single collation array elements get the data from the MVM
UCD database created with ucd2c.pl Any codepoints which have more than one
collation array element or if it is a sequence of codepoints, that uses the data
in src/strings/unicode_uca.c

Tangut, Ideographs, Nushu and Unassigned codepoints have collation values which
are generated algorithmically based on their codepoint. Hangul characters
decompose before they are pushed onto the array.

Like before, we support the ability to configure collation so you can reverse or
disable levels as you wish. The trick to this is knowing that for all collation
values: tertiary < secondary < primary

We use `level_eval_settings` to store the settings for each level, which we set
up based on the bitmask of the collation_mode argument to the fuction. If the
two levels are the same we are able to compare them based on the setting. If the
levels are not equal, we do not need to do this, since tertiary < secondary <
primary for all values.

Some info on our collation values. They are all 1 higher than those listed for
DUCET (Default Unicode Collation Element Table). The reason for this is that a 0
counts as 0 while a 1 is skipped and ignorable. This corresponds to things
listed as 0 in DUCET, which our implementation gives a value of 1. We only use 0
for the tertiary value of the level separator to ensure that longer strings win
(though we also have a fallback to ensure this happens in certain cases which
this isn't enough).