Fix ignoremark with Prepend characters and ordbaseat op
Previously ordbaseat would return the first codepoint in the
Synthetic instead of the correct base character. Now, when we
add a Synthetic, we check if the first codepoint is Prepend, and
if so we keep going until we find the last Prepend character.

This builds on the recent change so that synthetics are a single
array instead of storing the base codepoint separately (3bd371f1).