For degenerate Synth's with Prepend and Extend set base cp to 1st cp
If we have a degenerate that starts with Prepend and follows with Extend,
this commit sets it so that its base character will be the first one,
instead of the one following the last Prepend. I also set up a cache so we
don't have to repeatedly get the property values repeatedly in case the user
has the same codepoint repeating a large number of times.

Caching the codepoint was not strictly required for implementing the
Prepend/Extend degenerates change, though since we needed `cached_GCB`
to allow checking if the codepoint is Extend if the Prepend check fails
it was a simple addition.

These extra variables are only initialized and use if it encounters a Prepend
character at all. Since they are rare, this should not have overhead in mostly
cases except for one property value check.