Use the FSA for VMArrays
Just uses MVM_fixed_size_* functions where possible, based on a new
option added to the MVMCollectable's header flags. Using the FSA is a
stop toward making them not segfault when written to by multi-threaded
code. In addition, there are potential speedups for small arrays where
the storage can fit in one of the FSA's bin sizes, so if there's a free
slot in the bin allocation is much cheaper.