Just copy strands in MVM_string_join if able
When everything to be joined are strands, just copy them directly
instead of creating a single flat string. For small strings the
overhead of being strands counteracts the saving in copying, so leave
them them in the flat processing path, but for large strings it's much
faster to copy the strands.

For example:
  my $b = "bbbbbbbb" x 200;
  my $c = "cccccccccccccc" x 300;
  for ^100_000 {
      $ = join(",", $b, $c, "dddd")
  }
took ~3.7s before this change and ~1.5s after.

Also add NFG_CHECK and MVMROOT it and the STRAND_CHECK.

Also fix error in NFG_checker (since it's behind an #if, it only shows
up when MVM_DEBUG_NFG is enabled).