Download all emoji files and generate UCD for emoji 4+5
Since emoji names are not guaranteed to be unchanged, we now
download past emoji versions as well to ensure we don't break
compatibility in the future. This isn't expected to cause any
issues, just a few extra entries. For now the only thing that
may have changed is "man in business suit levitating: skin tone"
is now "man in suit levitating: skin tone".