Make thread GC nurseries scale to need
Many threads (such as the specialization worker and the Rakudo GC
supervisor) run and allocate relatively little. Thus they will never
fill a nursery, and so there's no point them having a sizable one.

This change starts all but the main thread off with a smaller nursery.
When a thread fills its nursery and triggers GC, then we treat that as
a sign that it'd like a larger one and double it size, up to the
nursery size limit. Any thread that never triggers GC in its life will
be left with just a 128KB nursery fromspace/tospace, meaning it will
have only 256KB of memory overhead for nursery tospace/fromspace as
opposed to 8MB.

This also means there's less memory to zero when doing the swapover
for some threads too, which means as well as a memory saving this
gives some CPU savings too.