Set frame->static_info in allocate_frame
So it's there right from the start. If the spesh log entry fills up
the log and causes it to be sent, then we might have a frame with no
static_info on the temp roots stack. This can cause a NULL pointer
dereference inside the GC. We could NULL-guard it there, but it's
less work to just make sure that situation can never happen.