pow_I can be very expensive, so GC-block thread
We may want to tune this based on exponent size, and may also want to
review if other bigint ops can be so very costly and do a similar
thing. In the meantime, this is enough to address the hang in the
power.t spectest after the resizable nursery changes.