Ensure ->caller/->static_info can't get outdated
Recent changes moved ->caller and ->static_info to be set up earlier,
to make sure a frame has them. However, in the case that we have a
frame allocated on the callstack rather than the heap, then MVMROOT
of that frame does nothing. This is normally fine; a frame not on the
heap will be hanging off something's `tc->cur_frame` and be marked if
GC is triggered. But in this particular case, we're making a new
frame, and it's not yet in that chain, so the two could go unmarked
and thus become outdated when the spesh log got sent.