use the FSA for MVMHash and friends
it's supposed to make concurrent hash accesses not as crashy
by using the "free at safepoint" mechanism, but more changes
are needed to make it any safer than it was before.

This commit creates a bunch of versions of the HASH_ functions
that end in _FSA and take a tc argument and then use the
fixed sized versions of alloc and free.