JIT [expr] process annotations on PHI node
Presumably due to code elimination, (frame handler) annotations can
sometimes end up on PHI nodes. Previously the expr JIT did not process
these because the iterator skipped such nodes. That was not such a
good idea as it seems to have led to a regression in Net::HTTP, which
was fortunately caught by integration tests.