Fix seekability detection on MacOS
POSIX does not define which devices must support lseek() and on
MacOS it returns some non-zero value that isn't -1 indicating failure
when lseek()ing TTY STDIN. This results in such handles marked
as seekable and in turn causes issues, such as mvm_eof() trying
to figure out whether we've reached EOF and falsely returning True
even when we didn't because the statbuf.st_size is zero.

After discussion[^1][^4] and some research, the solution in this
patch is what I came up with and that actually works. We try to seek
a handle with lseek and on MacOS, if it doesn't error out, then
try to see if it's a tty and claim non-seakability if it is.

Does this work? It passes the rakudo test[^1] to cover the bug,
but so far I'm having trouble completing spectest on the macos VM
I got (maybe something with VM I dunno).

Is this sane? I've no idea. I'm a n00b at this.

[1] (Notes section) http://man7.org/linux/man-pages/man2/lseek.2.html
[2] https://irclog.perlgeek.de/perl6-dev/2017-10-23#i_15341511
[3] https://github.com/perl6/roast/commit/27833272b4
[4] https://github.com/MoarVM/MoarVM/pull/731#issuecomment-338890688