Fix JITed nativeinvoke accessing wrong arg registers
The code blindly assumed that the native function's arguments are in the WORK
register corresponding to the argument number which worked by conincident.
Changes in the generated function body can thwart this assumption leading to
wrong values getting passed to the native function.

Fix by using the information from arg_ins instead.