JIT compile native calls with integer rw arguments
Passing a pointer to an integer instead of the integer itself to the native
function is easy. Getting the changed value back to the HLL code is the hard
part. Since the native function writes to the register that's passed in the
args buffer, the HLL code needs to read the value back from the args buffer.
That's what the new getarg_* ops are for.