Fix HLL representations of CArray elements out of sync after native calls
For CArrays with elements that are wrapped in HLL objects like strings,
CPointers or CArrays, we have to refresh the HLL representations after passing
those arrays to native functions as the native code may change the data. For
now just detect those cases and don't JIT compile the calls.
Fixes rakudo GH #1220