collapse_strands with memcpy if all strands are same type 4x faster
If all the strands to collapse are of the same type (ASCII, 8bit, or
32bit) then use memcpy to collapse the strands. If they are not all the
same type then we use the traditional grapheme iterator based collapsing
that we previously used to collapse strands.

If it's 8bit and a repetition with only one grapheme, it will use memset
to more quickly write the memory.

This is 4-4.5x faster as long as all the strands are of the same type.