ucd2c.pl: Add support for integer based enum properties
This acts similarly to our string enums except it uses an enum of
integers. Currently our integer properties store the integer values
themselves in the bitfield.

This change also makes improvements in property checking. Checking that
property value codes are not repeated and all of them are in the same
order. For now there is an exception for Numeric_Type because for
whatever reason Digit is removed and it is registered as its own
property.

This is a long standing bug where <:Digit> matches anything, but I am
leaving that as it is until another commit.