Fix bug causing breakage of Numeric_Type=Digit
This caused <:Digit> to match anything with regex, as well as codepoints
that should have had the Digit property value for Numeric_Type returning
an empty string instead.

This was caused by some code trying to add Digit as a subtype of
Numeric_Type for unknown reasons, causing it to be broken.