Re-work lexical handler lookup with inlines
The previous mechanism was a total cheat, and only survived because
marking lastexpayload as :noinline meant we never ran into any of the
cases that it didn't handle. This also prevented a lot of desirable
inlinings from taking place. This corrects the handling of lexical
handler resolution to work with inlines.