[JIT] Don't take a pointer-pointer to a moving object
This would cause segmentation faults because the alc->values array
(which backed the spillee object) could move in a realloc, and the
pointer-pointer to the ValueRef would point inside this object. We
don't actually need that, we can get by with two separate pointers
and a comma.

Note that we'd also assign to the (potentially moved) spillee object,
which is what I suspect was the real cause to the SEGV, but ASAN already
found the invalid read earlier.