ucd2c.pl: More work so codepoints are consistently allocated
Set $FIRST_POINT in set_next_points() instead of UnicodeData().
In get_next_point() set $POINTS_BY_CODE instead of in UnicodeData(). Add
an argument so we can make sure it only does this when called from
UnicodeData(), since there are issues adding certain codepoints which
then causes the database to for the most part break (the reason is not
clear).