Fix the Unicode database from adding int properties to str function
This was causing a compiler warning because of the type difference. We
shouldn't be requesting, or at least should expect to get an empty
string if we query an int property using the getuniprop_str op so there
should be no difference in Perl6 functionality since it only requests
using getuniprop_int.

Fixes issue #764