Fix crash in TTY flushing on WSL (Bash on Win10)
What little info I found seems to indicate there's a problem with
how flushing is implemented on WSL in that the error code for TTYs is
reported as errno 5 instead of 6 (which is what FlushFileBuffers returns
on normal Windows).

Fix by detecting WSL and not considering errno 5 a failure if the handle
is not seekable.

Fixes RT#131700: https://rt.perl.org/Ticket/Display.html?id=131700

[1] https://github.com/borgbackup/borg/issues/1961#issuecomment-335653560
[2] https://github.com/borgbackup/borg/issues/1961#issuecomment-338323934