Make log facts tolerate the odd outlier
If a type shows up less than one time in a hundred, then just go on
with the specialization, and let deopt handle the rare path. This nets
us a win in cases like `for ^10000 { @a[$_] = 0 }`, where the first
time in `AT-POS` we have no $!reified buffer, but every time after
that we do. (Note this will adapt naturally to whether program tends
to initialize lots of arrays and do little with them, or initialize
few arrays and do lots with them.)