make suspend/resume success report async
now the thread sends its own messages once it reacts.
to facilitate this, store a list of open requests in
the debug server data struct and go through it whenever
a thread resumes or suspends. ResumeAll and SuspendAll
requests get a counter field so that only the last
thread that resumed/suspended will report success.