Remove bogus comment and known value fact tweak
The comment seems wrong or perhaps just outdated because the code
got fixed; the temp registers play fine with SSA form and there is
still one use of the original register and one write to the final
destination one, so all's good. The known value fact update code,
on the other hand, was rather dubious; it should probably have been
looking at the temp facts and setting the res facts. However, with an
upcoming not_i optimization about to be added, that will take care of
things more elegantly, so just remove it.