Fix off-by-one error in uninline
The INLINE_END annotation sits on the last instruction of a chain of inlined
basic blocks, so it's this instruction's offset in the bytecode we record as
the end of the inlined code (mirroring the start offset).
This means that if we are exactly at cand->inlines[i].end, we are still in
inlined code and have to re-create the frame on uninlining.

This probably didn't cause issues before because inlined code always ends
with a goto.