Implement tile for OR operator
nine++ tried to implement the 'isnull' opcode as an expression
template as

  (or (zr $1) (eq $1 (^vmnull)))

Which should be totally valid, but didn't work because the OR operator
had no tile to implement. (There are checks for this to explicitly tell
you, but these seem to be optimized away). So this adds the simplest
possible tile to implement it.

Also fix the 'flagval' to suport 'NZ' and 'ZR', which are just the same
's EQ' and 'NE'.