Fix MVM_spesh_log_decont not logging anything
The condition in MVM_spesh_log_decont checks if we are still in the same place
in the bytecode as before calling ->fetch on the container. In other words, it
ensures that ->fetch did not invoke something (which it may). However, cur_op
has advanced 4 bytes since we saved it to prev_op in MVM_interp_run. Therefore
we have to add 4 bytes to prev_op when comparing it to cur_op.