Implement unmanaged_size in MVMSpeshLog repr
This means the GC understands the amount of space it really takes, and
so can trigger a full collection in a far more timely manner if we are
doing nothing but accumulating spesh logs (why that happens is another
issue, however). With this, the "leak" reported in Rakudo #1513 does at
least reach an upper boundary and stop growing. Prior to this, since
only the directly allocated memory of the spesh log was accounted for,
it would have taken a very long time for the GC to decide enough had
been promoted into gen2 to do a full collection (long enough for the
memory use to grow giant).