Add an nqp::tryfindmethod
Which evaluates to nqp::null in the event that the method cannot be
found. This will allow things that do nqp::can + nqp::findmethod to
avoid the double-resolution. While sometimes spesh can do that, this
can give better baseline performance before optimization and avoid a
double-lookup in the unoptimizable mega-morphic case.