getenvhash is supposed to :usehll
so make sure it only returns objects of the
right type, even if there's an object in
the cache already.

The slurpy_hash_type is the same in NQP and
rakudo as far as I can see, so this shouldn't
actually trigger, but it's probably better to
be safe here.