Throw out repeated/redundant getspeshslot calls
if the target register and the spesh slot match
and there's no instruction in between two calls
that write to the same register, we can drop the
later getspeshslot call.

Looking at !cursor_start for example showed code
with multiple native attribute binds in a row would
grab the type object for binding multiple times
in a row. They were only being kept alive for deopt
reasons, I believe, but for now this ought to be
a win.