[JIT] Release only spill slots without active overlap
When a value spilled to memory has ceased to be useful, its memory
slot can be reused. However, it is possible that there exists an
active live range that starts before the end of the spilled values'
active range. If the memory slot is released it might then be possible
that this overlapping live range might need to be spilled itself and
select the same slot, causing the original value to be overwritten in
the region where the ranges overlap.

To prevent this, compute the start of the earliest active live
range (bounded by the current tile order number for the case when
there are no active live ranges) to determine which of the spilled
value slots we can reuse.