Create store_write_barrier macro
This orders the STORE node before the write-barrier, so that we don't
need to spill the computed address over the conditional call to
MVM_gc_write_barrier_hit. It is not necessary to do the write barrier
prior to the actual assignment (its purpose is to register the root
object as a cross-generation root if the root is old and the value
object is new, and this can happen at any time before the next GC
run).

I kind of want to treat store-with-write-barrier as a single, uniform
abstraction, but it is too big in practice to be implemented as an
operator. That would also imply that there also had to be a
store_write_barrier tile, and that the linear scan allocator had to
treat that tile it with register constraints similar to CALL. Encoding
such per-tile constraints and satisfying them in a general way is not
yet implemented, and I'm not eager to start doing so. So instead I
will try and 'export' this macro to the JIT runtime as a template.