[JIT] Warn for things that should be a size but may not be
In MasterDuke++'s set of templates there is a (CALL) node in the
existskey template that indicates its size as 'int'. The 'int' keyword
is however resered to indicate a *type* in a (CARG) node - what was
meant was the int_sz keyword instead.

This is confusing but somewhat difficult to clean up entirely. What we
can do easily though is warn when an argument that *should* be a size
looks like it isn't (not a number, not a macro, and doesn't end with
_sz).