Make sure to throw in shiftjis decode if byte after lead
If Shift_JIS_lead is set at the end of decode throw since we should have
seen a byte following it.

Don't throw in decodestream if we have no byte after lead, since we may
resume the stream, and last_accept_pos is properly set so the lead byte
should be re-read when starting the stream again. Currently Perl 6
doesn't throw if we have a lead byte at the end of the file when
slurping it. This is also a bug that exists with the utf8 decoder so
this may need to be dealt with on the Rakudo side of things.