Fix windows1252/1 decodestream replacement with last pos repl.
If the last position byte is bad and there is a replacement, we would
exit the while loop since we had moved past the last position. Ensure
that the loop continues until we write the full replacement. This bug
only occured with replacements longer than one character.

Example: "abcX" where X is the bad character and replacement = "123"
would result in "abc1" instead of the correct result of "abc123"